@@ -1,6 +1,6 @@
 package com.example.pedroimai.kotlinrx2.movie.listing
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
